admin/backend: Do not clear write-in candidates along with cvrs in qualified mode

+++ apps/admin/backend/src/store.ts
@@ -1901,13 +1901,15 @@ export class Store implements BaseStore {
           )
         `
       );
-      this.client.run(
-        `
-          delete from write_in_candidates
-          where election_id = ?
-        `,
-        electionId
-      );
+      if (!this.getSystemSettings(electionId).areWriteInCandidatesQualified) {
+        this.client.run(
+          `
+            delete from write_in_candidates
+            where election_id = ?
+          `,
+          electionId
+        );
+      }
       this.deleteEmptyScannerBatches(electionId);
     });
